python/src/auth/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.2
